Move music Spotify implementation to subpackage within music.
